xscreensaver disabled

+++ cma_via_ssh/install_cma.sh
@@ -1,13 +1,15 @@
 #!/bin/bash
 
-read tmpPath <<<$@
+read tmpPath appDir appName <<<$@
 echo -e "\tApplication release archive: $tmpPath"
+echo -e "\tApplication instalation path: $appDir"
+echo -e "\tApplication name: $appName"
 # echo -e "\tRoot password: $rPassword"
+
 
 isInstalled() {
   target=$1
   isInstalledResult=$(dpkg-query -l $target)
-
   if [[ $isInstalledResult ]]; then
       echo -e "$target is allready installed"
      return 0
@@ -19,18 +21,26 @@ isInstalled() {
 
 packageName="unzip"
 if isInstalled $packageName; then
-    echo
+    :
 else
-    echo -e "\n${BLUE}Installing UnZip on remote $hostname...${NC}"
+    echo -e "\t${BLUE}Installing UnZip on remote $hostname...${NC}"
     su -c 'apt update'
     su -c 'apt install unzip -y'
 fi 
 
-echo -e "\n${BLUE}Extracting files on remote $hostname...${NC}"
+echo -e "\t${BLUE}Extracting $tmpPath on remote $hostname...${NC}"
 rm -rf '/tmp/release'
 unzip $tmpPath -d '/tmp/release/'
-echo -e "\n${BLUE}Coping files on remote $hostname...${NC}"
-rm -rf '/home/scada/app/cma/'
-mkdir -p /home/scada/app
-cp -ra "$(find /tmp/release -name crane_monitoring_app -type f -printf '%h' -quit)/" /home/scada/app/cma/
-chmod +x '/home/scada/app/cma/crane_monitoring_app'
+extractedDir="$(find /tmp/release -name $appName -type f -printf '%h' -quit)/"
+if [ -d "$extractedDir" ]; then
+    echo -e "\textracted to "$extractedDir""
+    echo -e "\t${BLUE}Coping files on remote $hostname...${NC}"
+    rm -rf '/home/scada/app/cma/'
+    mkdir -p /home/scada/app
+    cp -ra $extractedDir $appDir
+    chmod +x "$appDir/$appName"
+    echo -e "\tInstalled successful to "$appDir""
+else
+    echo -e "\textracted not found"
+    echo -e "\tNot installed"
+fi
